+++ src/main/java/com/bashilya/blog/comment/service/CommentApiService.java
@@ -3,8 +3,12 @@ package com.bashilya.blog.comment.service;
 import com.bashilya.blog.article.api.response.ArticleResponse;
 import com.bashilya.blog.article.exception.ArticleNotExistException;
 import com.bashilya.blog.article.repository.ArticleRepository;
+import com.bashilya.blog.auth.exceptions.AuthException;
+import com.bashilya.blog.auth.exceptions.NotAccessException;
+import com.bashilya.blog.auth.service.AuthService;
 import com.bashilya.blog.base.api.request.SearchRequest;
 import com.bashilya.blog.base.api.response.SearchResponse;
+import com.bashilya.blog.base.service.CheckAccess;
 import com.bashilya.blog.comment.api.request.CommentRequest;
 import com.bashilya.blog.comment.api.request.CommentSearchRequest;
 import com.bashilya.blog.comment.mapping.CommentMapping;
@@ -13,9 +17,11 @@ import com.bashilya.blog.comment.exception.CommentNotExistException;
 import com.bashilya.blog.comment.model.CommentDoc;
 import com.bashilya.blog.comment.repository.CommentRepository;
 import com.bashilya.blog.user.exception.UserNotExistException;
+import com.bashilya.blog.user.model.UserDoc;
 import com.bashilya.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
@@ -29,22 +35,21 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class CommentApiService {
+public class CommentApiService extends CheckAccess<CommentDoc> {
     private final CommentRepository commentRepository;
     private  final MongoTemplate mongoTemplate;
     private final UserRepository userRepository;
     private final ArticleRepository articleRepository;
+    private final AuthService authService;
 
-    public CommentDoc create(CommentRequest request) throws CommentExistException, UserNotExistException, ArticleNotExistException {
+    public CommentDoc create(CommentRequest request) throws AuthException, ArticleNotExistException {
 
-        if (userRepository.findById(request.getUserId()).isPresent() == false) {
-            throw new UserNotExistException();
-        }
+        UserDoc userDoc = authService.currentUser();
         if (articleRepository.findById(request.getArticleId()).isPresent() == false) {
             throw new ArticleNotExistException();
         }
 
-        CommentDoc commentDoc = CommentMapping.getInstance().getRequestMapping().convert(request);
+        CommentDoc commentDoc = CommentMapping.getInstance().getRequestMapping().convert(request,userDoc.getId());
 
         commentRepository.save(commentDoc);
 
@@ -91,15 +96,16 @@ public class CommentApiService {
     }
 
 
-    public CommentDoc update(CommentRequest request) throws CommentNotExistException {
+    public CommentDoc update(CommentRequest request) throws CommentNotExistException, NotAccessException, AuthException {
         Optional<CommentDoc> commentDocOptional = commentRepository.findById(request.getId());
         if (commentDocOptional == null) {
             throw new CommentNotExistException();
         }
 
         CommentDoc oldDoc = commentDocOptional.get();
+        UserDoc owner = checkAccess(oldDoc);
 
-        CommentDoc commentDoc = CommentMapping.getInstance().getRequestMapping().convert(request);
+        CommentDoc commentDoc = CommentMapping.getInstance().getRequestMapping().convert(request,owner.getId());
         commentDoc.setId(request.getId());
         commentDoc.setArticleId(oldDoc.getArticleId());
         commentDoc.setUserId(oldDoc.getUserId());
@@ -109,7 +115,18 @@ public class CommentApiService {
         return commentDoc;
     }
 
-    public void delete(ObjectId id) {
+    public void delete(ObjectId id) throws NotAccessException, AuthException, ChangeSetPersister.NotFoundException {
+        checkAccess(commentRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         commentRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(CommentDoc entity) {
+        return entity.getUserId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
