+++ src/main/java/com/bashilya/blog/todoTask/service/TodoTaskApiService.java
@@ -1,7 +1,11 @@
 package com.bashilya.blog.todoTask.service;
 
+import com.bashilya.blog.auth.exceptions.AuthException;
+import com.bashilya.blog.auth.exceptions.NotAccessException;
+import com.bashilya.blog.auth.service.AuthService;
 import com.bashilya.blog.base.api.request.SearchRequest;
 import com.bashilya.blog.base.api.response.SearchResponse;
+import com.bashilya.blog.base.service.CheckAccess;
 import com.bashilya.blog.todoTask.api.request.TodoTaskRequest;
 import com.bashilya.blog.todoTask.api.request.TodoTaskSearchRequest;
 import com.bashilya.blog.todoTask.mapping.TodoTaskMapping;
@@ -10,9 +14,12 @@ import com.bashilya.blog.todoTask.exception.TodoTaskNotExistException;
 import com.bashilya.blog.todoTask.model.TodoTaskDoc;
 import com.bashilya.blog.todoTask.repository.TodoTaskRepository;
 import com.bashilya.blog.user.exception.UserNotExistException;
+import com.bashilya.blog.user.model.UserDoc;
 import com.bashilya.blog.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.apache.catalina.User;
 import org.bson.types.ObjectId;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
@@ -28,18 +35,17 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-public class TodoTaskApiService {
+public class TodoTaskApiService extends CheckAccess<TodoTaskDoc> {
     private final TodoTaskRepository todoTaskRepository;
     private  final MongoTemplate mongoTemplate;
     private final UserRepository userRepository;
+    private final AuthService authService;
 
-    public TodoTaskDoc create(TodoTaskRequest request) throws TodoTaskExistException, UserNotExistException {
+    public TodoTaskDoc create(TodoTaskRequest request) throws AuthException {
 
-        if (userRepository.findById(request.getOwnerId()).isEmpty()) {
-            throw new UserNotExistException();
-        }
+        UserDoc userDoc = authService.currentUser();
 
-        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request);
+        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request,userDoc.getId());
 
         todoTaskRepository.save(todoTaskDoc);
 
@@ -50,11 +56,11 @@ public class TodoTaskApiService {
         return todoTaskRepository.findById(id);
     }
 
-    public SearchResponse<TodoTaskDoc> search(TodoTaskSearchRequest request) {
+    public SearchResponse<TodoTaskDoc> search(TodoTaskSearchRequest request) throws AuthException {
 
-        if (request.getOwnerId() == null) throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        UserDoc userDoc = authService.currentUser();
 
-        Criteria criteria = Criteria.where("ownerId").is(request.getOwnerId());
+        Criteria criteria = Criteria.where("ownerId").is(userDoc.getId());
         if (request.getQuery() != null && request.getQuery() != "") {
             criteria = criteria.orOperator(
 
@@ -77,15 +83,16 @@ public class TodoTaskApiService {
     }
 
 
-    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNotExistException {
+    public TodoTaskDoc update(TodoTaskRequest request) throws TodoTaskNotExistException, NotAccessException, AuthException {
         Optional<TodoTaskDoc> todoTaskDocOptional = todoTaskRepository.findById(request.getId());
         if (todoTaskDocOptional == null) {
             throw new TodoTaskNotExistException();
         }
 
         TodoTaskDoc oldDoc = todoTaskDocOptional.get();
+        UserDoc userDoc = checkAccess(oldDoc);
 
-        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request);
+        TodoTaskDoc todoTaskDoc = TodoTaskMapping.getInstance().getRequestMapping().convert(request,userDoc.getId());
         todoTaskDoc.setId(request.getId());
         todoTaskDoc.setOwnerId(oldDoc.getOwnerId());
         todoTaskRepository.save(todoTaskDoc);
@@ -93,7 +100,18 @@ public class TodoTaskApiService {
         return todoTaskDoc;
     }
 
-    public void delete(ObjectId id) {
+    public void delete(ObjectId id) throws NotAccessException, AuthException, ChangeSetPersister.NotFoundException {
+        checkAccess(todoTaskRepository.findById(id).orElseThrow(ChangeSetPersister.NotFoundException::new));
         todoTaskRepository.deleteById(id);
+    }
+
+    @Override
+    protected ObjectId getOwnerFromEntity(TodoTaskDoc entity) {
+        return entity.getOwnerId();
+    }
+
+    @Override
+    protected AuthService authService() {
+        return authService;
     }
 }
